requirements.txt
Django
djangorestframework
requests
django-jsonfield
django-filter
django-cors-headers